📃 chore(prod): allow all kinds of catalog
By default the catalog will only allow ingestion of entities with the kind `Component`, `API` and `Location`. In order to allow entities of other kinds to be added, you need to add rules to the catalog. Rules are added either in a separate `catalog.rules` key, or added to statically configured locations.